Move ParserGenerator to separated class from Workflow
